Simplify the use of params.pp pattern
With recent discussions about data in modules going on, it has occurred
to me that I was not doing things quite as simply as I could. There was
a lot of additional logic in the module just to handle passed params vs
default params.

This commit simplifies the params approach and remove quite a few lines
of code simplifying things.