Remove useless filename selection
This module has the ability to set the path for the location of the
nsswitch.conf file based on the Operation System the manifest is being
compiled for. This is a useless feature. After some more than simple
cursory research, I found out I was simply over anticipated platform
differences for the NSS libs. The NSS libraries use the same
configuration location for nsswitch.conf on every *nix that utilizes
NSS.

This commit removes the useless code and hard-codes the location. Thus
simplifying the params class, and removing needless variable lookups in
Puppet (hooray insignificant performance increase). This commit also
explicitly specifies the ensure value for the file resource to be in
better alignment with the Puppet Labs style guide rules.