Convert to metadata.json
The Modulefile format is not deprecated and will soon no longer be
supported. This commit converts to the metadata.json format. The
metadata.json format has the advantage of being easier to check
syntactically and the new format supports a much wider range of data.
This conversion is taking advantage of the newly supported data and
adding some additional much needed pieces of data.