Fix fact name in spec tests
In the spec tests, the final test that ensures that the code fails on unknown
platforms had a typo. This typo was with the artificial setting of the
'operatingsystem' fact. This typo kept the unit test from getting an official
value. It is only by luck that the test had been working.

This commit corrects the typo ensuring the test will work successfully.