Remove Gemfile.lock from the .gitignore
I've learned a thing or two about spec testing and the use of bundler in a ruby
project in general. The biggest misconception I had was about that on the
'Gemfile.lock' file and it's purpose. I now realize that the 'Gemfile' guides
Bundler to build the env, but Gemfile.lock truely tracks every aspect of the
external deps. I foolishly was ignoring this file which ensured that, since I
did not have locked gem version, every build of the environment for test might
be different.

This commit starts managing the 'Gemfile.lock' to ensure that automated testing,
as well as testing for contributers works the same across the board. Another
aspect of the project to manage, sure, but an important one I feel.