Refactor unit tests for changes in Rspec/Rspec-Puppet
With Rspec 3.x and Rspec-Puppet 2.x the method of testing that Puppet code
raises an error needs to change a bit. While I'm not 100% sure what changed
inside Rspec, the result is that the example in the rspec-puppet documentation
is no longer working. It instead causes an internal Rspec expectation error
to be thrown and makes the test fail.

This commit turns to an alternate form for the test. The new form simply lets
rspec-puppet run the catalog compilation and then catches the error as it's
thrown. I have verified that this fix is functional and will not return a false
positive. I do however need to better articulate why in Rspec terms why this
happened and why this fix works.