Update spec tests to use Rspec 3.x syntax
Starting in Rspec 3.x, the syntax around one-liners has evolved. The `should`
syntax, while still available, is no longer monkey patched on Object. It is now
part of the expectation system technically. While it still works similar to the
syntax of Rspec's past, it is suggested that the syntax be changed to the more
informative and descriptive `is_expected.to` syntax.

This commit updated the spec test syntax.