Removing Gemfile.lock - I was wrong
So, I think I definitely understand the Gemfile.lock a lot more now, and I also
understand that I should not include in this project. If I have gems I need to
keep locked in place, the Gemfile is where I will need to do it. This is due to
the fact that I'm using a build matrix based on the Puppet gem version and need
to have Bundler recalculate the gem dependencies makes the Gemfile.lock mostly
useless. Oh well, live and learn.

This commit removes once again the Gemfile.lock as well as ensures certain truths
about the world for this project in terms of rspec and rspec-puppet.