Fixup the travis-ci configuration
This commit corrects the travis-ci configuration. Things have changed in the
world sine I've last done a serious amount of work here. For example the
Puppet FOSS Gem 2.7.0 has been yanked. This commit updates the testing to no
longer include testing of FOSS 2.7 and as such this module is no longer
considered supported on the Puppet 2.7 series.

This commit also updates the testing to be performed against the most recent
releases of Puppet 3 and Puppet 4 that are available. Since Puppet 3 and 4
are both semantically versioned, this method of testing should be sufficient.