Refactor for Puppet 4
Puppet 4, a game changer in ways. Puppet 4 introduces a new type system that
makes validation of inputs a more natural part of the system. One of the
missing abilities of the validation methods in stdlib was the ability to
accept multiple kinds of input and accurately validate them without 50 lines of
code. This requirement was being half met with the `validate_multi()` function,
however a better method is always preferred. The new `Variant` abstract type
along with the ability to specify explicit data types means that what was
needed before is now truly a reality.

This commit remove usage of the `validate_mutli()` function by using the
`Variant` feature of Puppet 4 to maintain the previous behavior.