Extend Rspec testing to validate input types
The previous set of tests did not provide good coverage of the behaviors around
input for the `nsswitch` class. It simple ran with default values to ensure
that a file could be generated for each supported system.

This commit adds tests to validate the ability to provide input values of two
supported types, String and Array.