Remove undesirable tags from metadata
While any tags are valid tags, there are certain tags on the Puppet
Forge that are seen as undesirable. These tags are typically tags that
try to suggest what OS or distribution the module is geared for. The
reason tags representing this are considered undesirable id due to the
metadata.json's very explicit compatibility fields. These should be
preferred over tags.

This commit removes tags that conflict potentially with the purpose of
the compatibility field in the metadata.json file.