Add explicit support for EL 7
There have been some minor updates to the defaults in the `nsswitch.conf`
file for EL 7 based systems. This requires that the module account for the
differences as to maintain the level of default operation originally
committed to.

This commit includes defaults for both EL6 and EL7 based systems. These
differences are selected using the `operatingsystemmajrelease` fact that
was added to Facter in the 1.7.0 release. This change means that this
module will no longer function on versions of Facter less than 1.7.0
when EL systems are being managed.