Add support for CloudLinux distributions
After testing with two fresh installs from distributor ISO images (6.6
and 7.0) it appears that CloudLinux conforms to the standard mold of EL
systems. As such, this module already has the proper defaults for the
`nsswitch.conf` file on this distribution.

This commit adds CloudLinux as an acceptable operating system. This
commit also adds CloudLinux to the spec testing.