Move from ERB to EPP templates
With Puppet 4, a new template language is now available, the EPP
template (as in "embedded puppet"). This new template language has a few
advantages including better clarity. Overall it does not look largely
different from the ERB days, but since this module it looking to be a
prime example of Puppet 4 greatness, it makes sense to move to EPP.

This commit removes the ERB template, replacing it with a well formed
EPP template. The template call in the `nsswitch` class is updates to
use EPP. The requirement for `puppetlabs/stdlib` is being re-added as
the EPP requires the "join" function from the library.