Use numeric UID and GID
Without this commit, this module will attempt to use "root" as the name
of the user and group to own the configuration to. This is not wise
given the range of systems this module will eventually come to support.
Using a numberic UID/GID for ownership is far safer since the super user
is always "0" regardless of human readable name.

This commit uses the UID/GID of "0" for the ownership of the
`nsswitch.conf` file.