Add LinuxMint aupport
This commit adds support for LinuxMint 17.2 release. The default section
was provided via an issue open of GitHub. The submitter, @HendryLeo,
only mentioned 17.2 in the issue and thus this module will only commit
to support that version. It is important to know that at this time, I
have not directly tested this on a LinuxMint 17.2 system.