Adding `initgroups` support
Without this commit, there was no way to control the `initgroups`
setting using this module.

This commit adds support for the `initgroups` setting. All previous
research suggests there is no platform supported by this module that
has a default for this setting. If there is and I've missed it,
hopefully someone will tell me.